backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3

# For filtering support in DRF viewsets
django-filter>=24.1


